+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+Added
+
+  - `default` optional parameter to the `get()` method of `ConfigFile`. 
+    The allows a default value to be fallen back to if the given key is missing.
+
 ## 0.8.0 - 2020-02-02
 
 Added